Add tests for uploadAndPublish
We verify that result expected is returned and that commands we wanted to be were invoked.

There are two goals in mind,
1) Test that commands behave the way we think they should
2) Be confident that we will be able to re-verify this from within sbt later

With the latter in mind these tests also write the commands performed to disk
and verify that they match what we think they should. sbt-scripted tests should
then just be a formality (they are very slow to run)

Also as part of this commit
- first stab at fake free compiler interpreter that will be used in most of these tests
- Realised that in order to use our FakeAWSCompiler we will need to rejig sbt setting